+++ jetpack_chili/src/main/java/kg/devcats/compose/jetpack_chili/components/pin/PinStatusType.kt
@@ -1,7 +1,7 @@
 package kg.devcats.compose.jetpack_chili.components.pin
 
-enum class PinStatusType {
-    Error,
-    Success,
-    None
+sealed class PinStatusType {
+    data class Error(val repeat: Int = 3, val duration: Int = 50) : PinStatusType()
+    data class Success(val repeat: Int = 1, val duration: Int = 200) : PinStatusType()
+    data object None : PinStatusType()
 }
